feat: restructure UdpServer implementation; move to separate header and source files, update file paths
